+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- fix: make getBuilder optional #41 
 
 ## [0.10.0] - 2023-05-22
 ### Added